chore: readme changes, using kebab case for files and imports, proper indentation on type files
