+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "@zbd/node",
   "version": "0.6.2",
-  "description": "ZEBEDEE NodeJS SDK",
-  "author": "ZEBEDEE",
+  "description": "ZBD NodeJS SDK",
+  "author": "ZBD",
   "license": "MIT",
   "scripts": {
     "prebuild": "yarn run clean",
